% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bin-method.R
\name{binBased_DMA}
\alias{binBased_DMA}
\title{Bin-based method}
\usage{
binBased_DMA(
  peaks,
  pathToHTseqCount,
  sampleCondition,
  txDB = NA,
  condition1,
  condition2,
  binwidth = 0.3
)
}
\arguments{
\item{peaks}{A GRanges object which contains the single nucleotide peaks from the peak calling
result of pureCLIP.}

\item{pathToHTseqCount}{A path way to all the result of HTSeq-count folder.}

\item{sampleCondition}{A vector which contains the condition of experiment. For example,
sampleCondition <- c(rep("KO",3), rep("WT",3)).}

\item{txDB}{If the user has assigned the peaks to the gene, please keep txDB=NA and save the gene
ID of each peaks under the column "gene_id". If the user has not assigned the peaks yet, this
parameter should point to a txDB object of the annotation file.}

\item{condition1}{A vector which indicates the colname of the control or WT group in the input
peaks objects. For example, condition1 <- c("read_WT_1", "read_WT_2", "read_WT_3").}

\item{condition2}{A vector which indicates the colname of the KO or other treatment in the input
peaks objects. For example, condition2 <- c("read_KO_1", "read_KO_2", "read_KO_3").}

\item{binwidth}{The bin width of gene expression changes for grouping the peaks.}
}
\description{
A function to run the Bin-based differential methylation analysis.
}
\author{
You Zhou, Kathi Zarnack
}
